fix:fix some bug cause clang-tidy modified

+++ src/pika_zset.cc
@@ -859,7 +859,7 @@ void ZPopmaxCmd::DoInitial() {
     count_ = 1;
     return;
   }
-  if (pstd::string2int(argv_[2].data(), argv_[2].size(), static_cast<long long*>(&count_)) == 0) {
+  if (pstd::string2int(argv_[2].data(), argv_[2].size(), static_cast<long *>(&count_)) == 0) {
     res_.SetRes(CmdRes::kInvalidInt);
     return;
   }
@@ -893,7 +893,7 @@ void ZPopminCmd::DoInitial() {
     count_ = 1;
     return;
   }
-  if (pstd::string2int(argv_[2].data(), argv_[2].size(), static_cast<long long*>(&count_)) == 0) {
+  if (pstd::string2int(argv_[2].data(), argv_[2].size(), static_cast<long *>(&count_)) == 0) {
     res_.SetRes(CmdRes::kInvalidInt);
     return;
   }
